+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Please mark all change in change log and use the ticket from JIRA.
 - \#12 - Pure CPU version for Milvus
 - \#77 - Support table partition
 - \#226 - Experimental shards middleware for Milvus
+- \#346 - Support build index with multiple gpu
 
 ## Improvement
 - \#275 - Rename C++ SDK IndexType